Backend/citation-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai>=1.0.0
psycopg2-binary
elasticsearch
qdrant-client
neo4j
PyJWT
rich
httpx
